mod pre/footer: Alan card switch for crew card
