+++ app/Home/components/Footer/Footer.tsx
@@ -1,35 +1,45 @@
+'use client';
 import React from 'react';
-
-import alan from '@app/assets/img/alan-buscaglia.png';
-import Image from 'next/image';
 
 import Controller from './Controller';
 import styles from './footer.module.css';
+import FooterCard from './FooterCard';
+import team from '../PreFooter/team.json';
+export interface Member {
+  name: string;
+  pic: string;
+  role: string;
+  description: string;
+  linkedin: string;
+  portfolio: string;
+  github: string;
+  fb: string;
+  twitter: string;
+  twitch: string;
+  visibility: boolean;
+}
 
 const Footer = () => {
+  const [currentContent, setCurrentContent] = React.useState(3);
+  React.useEffect(() => {
+    const interval = setInterval(() => {
+      setCurrentContent((prevContent) => {
+        if (prevContent === team.length - 1) {
+          return 0;
+        } else {
+          return prevContent + 1;
+        }
+      });
+    }, 3000);
+    return () => clearInterval(interval);
+  }, [currentContent]);
+  const croppedTeam = team[currentContent];
   return (
     <footer className={styles.footer}>
       <hr className={styles.footer__divider} />
+      <h2 className={styles.footercard__title}>Gentleman Rockstars</h2>
       <div className={styles.footer__container}>
-        <article className={styles.footer__alanCard}>
-          <Image
-            className={styles.footer__alanImg}
-            src={alan}
-            alt="foto de Alan Buscaglia"
-          />
-          <section className={styles.footer__textContainer}>
-            <header className={styles.footer__alanTitle}>Alan Buscaglia</header>
-            <p>Ingeniero y arquitecto de Software Front-end</p>
-            <p>
-              ChatGPT dice que mi contenido es
-              <i> valioso y cumple con los estándares de calidad.</i>
-            </p>
-            <p>
-              <q>Desde mi cabeza, corazón y experiencia, para todo el mundo.</q>
-            </p>
-          </section>
-        </article>
-
+        <FooterCard key={'card' + croppedTeam.name} member={croppedTeam} />
         <section className={styles.footer__aboutCommunityContent}>
           <header>Sobre mi contenido</header>
           <p>
