+++ app/Home/components/Footer/footer.module.css
@@ -5,8 +5,21 @@
   background: linear-gradient(180deg, #000000 88%, #ea1889 100%);
   color: rgba(255, 255, 255, 0.9);
   font-size: 12px;
-  height: 15rem;
+  height: fit-content;
   margin: 4.7rem 0 0 -4.7rem;
+}
+.footercard__title {
+  font-family: "Fira Mono";
+  color: aliceblue;
+  display: flex;
+  height: -moz-fit-content;
+  height: fit-content;
+  font-size: 1.15rem;
+  font-weight: 500;
+  letter-spacing: 0.05em;
+  width: 100%;
+  margin-left: 4.5rem;
+  justify-content: flex-start;
 }
 .footer__container {
   display: grid;
@@ -16,47 +29,6 @@
   margin-left: 1.5em;
   align-items: start;
   align-content: center;
-}
-.footer__alanCard {
-  display: flex;
-  flex-direction: row;
-  padding: 1em;
-  border-radius: 3px 18px 3px 18px;
-  background-color: #ea1888b7;
-  transition: 250ms;
-}
-.footer__alanCard:hover {
-  transition: 250ms;
-  border: hsla(328, 83%, 51%, 0.85);
-  box-shadow: 0px 0px 6px #ea1889;
-  background-color: #ea1889;
-  border-radius: 3px;
-  z-index: 99999;
-}
-.footer__alanTitle {
-  font-weight: 600;
-  font-size: larger;
-  text-shadow: 0px 0px 2px #eeeeee;
-  text-align: center;
-}
-.footer__alanImg {
-  border-radius: 50%;
-  aspect-ratio: 1/1;
-  height: 120px;
-  width: 120px;
-}
-.footer__alanImg:hover {
-  transition: 250ms;
-  box-shadow: 0px 0px 6px #000000;
-}
-.footer__textContainer {
-  display: flex;
-  flex-direction: column;
-  padding: 0 0.5em 0 0.5em;
-}
-.footer__textContainer > p {
-  margin: 0.25em;
-  text-align: center;
 }
 .footer__aboutCommunityContent {
   text-align: center;
