--- app/Home/components/PreFooter/PrefooterCard.tsx
@@ -1,87 +1,0 @@
-'use client';
-
-import React from 'react';
-
-import { Member } from './Prefooter';
-import styles from './prefooter__card.module.css';
-import {
-  FacebookIcon,
-  GithubIcon,
-  LinkedInIcon,
-  TwitterIcon,
-} from '../Footer/Icons';
-
-type Props = {
-  member: Member;
-};
-
-const PrefooterCard = ({ member }: Props) => {
-  if (member.visibility === true)
-    return (
-      <article
-        key={member.name + ' article'}
-        className={styles.prefooter__card}
-      >
-        <div className={styles.prefooter__card_info}>
-          <section className={styles.prefooter__card_side}>
-            <img
-              alt={member.name + ' pic'}
-              className={styles.prefooter__card_pic}
-              src={member.pic}
-            />
-            <header className={styles.prefooter__card_header}>
-              <p className={styles.prefooter__card_name}>{member.name}</p>
-              <p>{member.role}</p>
-            </header>
-          </section>
-          <section className={styles.prefooter__card_description}>
-            <p>{member.description}</p>
-          </section>
-        </div>
-        <div className={styles.prefooter__card_info_socials}>
-          {member.fb !== '' ? (
-            <a
-              href={member.fb}
-              style={{ marginBottom: '0.15rem' }}
-              target="__blank"
-              title="Facebook"
-            >
-              <FacebookIcon />
-            </a>
-          ) : null}
-          {member.twitter !== '' ? (
-            <a
-              href={member.twitter}
-              style={{ marginTop: '0.08rem', marginRight: '0.25rem' }}
-              target="__blank"
-              title="Twitter"
-            >
-              <TwitterIcon />
-            </a>
-          ) : null}
-          {member.linkedin !== '' ? (
-            <a
-              href={member.linkedin}
-              style={{ marginBottom: '0.22rem' }}
-              target="__blank"
-              title="LinkedIn"
-            >
-              <LinkedInIcon />
-            </a>
-          ) : null}
-          {member.github !== '' ? (
-            <a
-              href={member.github}
-              style={{ marginTop: '0.12rem' }}
-              target="__blank"
-              title="GitHub"
-            >
-              <GithubIcon />
-            </a>
-          ) : null}
-        </div>
-      </article>
-    );
-};
-
-export default PrefooterCard;
