--- app/Home/components/PreFooter/prefooter__card.module.css
@@ -1,61 +1,0 @@
-.prefooter__card {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  border-top: 2px solid hotpink;
-  border-left: 2px solid hotpink;
-  border-right: 2px solid hotpink;
-  background-color: rgb(255, 245, 245);
-  border-radius: 10px;
-  width: 19rem;
-  padding: 0 1rem 0 1rem;
-  height: 17.5rem;
-  flex-direction: row;
-}
-.prefooter__card_side {
-  display: flex;
-  flex-direction: column;
-  padding: 0.75rem 0.5rem 0 0.5rem;
-}
-.prefooter__card_pic {
-  border-left: 2px solid #ea1889;
-  border-right: 2px solid #ea1889;
-  width: 6.5rem;
-  height: 6.5rem;
-  border-radius: 2px 16px;
-}
-.prefooter__card_header {
-  font-family: Cambria, Cochin, Georgia, Times, "Times New Roman", serif;
-}
-.prefooter__card_name {
-  font-size: 1.25rem;
-  font-weight: 600;
-  margin-bottom: -0.5rem;
-}
-.prefooter__card_info {
-  display: flex;
-  padding: 0 0.5rem 0 0.5rem;
-  text-align: center;
-  flex-direction: row;
-  align-items: center;
-  width: 100%;
-}
-.prefooter__card_description {
-  padding-top: 0.5rem;
-  font-size: 13px;
-  font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-    Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-  font-weight: 600;
-  word-wrap: break-word;
-}
-.prefooter__card_info_socials {
-  display: flex;
-  width: 80%;
-  justify-content: space-around;
-  align-items: center;
-  align-self: center;
-  padding-top: 0.3rem;
-  border-top: 2px solid #ea1889;
-  border-radius: 20px 20px 0 0;
-  background-color: black;
-}
